Address flake8 concerns
