+++ tests/test_config.py
@@ -1,7 +1,3 @@
-from os import environ
-
-import pytest
-
 from pymagicc.config import (
     ConfigStore,
     config,
@@ -9,50 +5,6 @@ from pymagicc.config import (
     lookup_defaults,
     lookup_env,
 )
-
-
-def temp_set_var(store):
-    """
-    Temporary sets a value of a Dict-like object for the duration of a test
-    :param store: A Dict-like object which holds key-value pairs. The store is
-        restored to its original state at the end of the test
-    """
-    prev_values = {}
-
-    def set_var(name, value):
-        if name not in prev_values:  # Only remember the first value
-            prev_values[name] = store.get(name)
-        if value is None:
-            try:
-                del store[name]
-            except KeyError:
-                pass
-        else:
-            store[name] = value
-
-    def cleanup():
-        # Clean up any set variables
-        for n in prev_values:
-            if prev_values[n] is not None:
-                store[n] = prev_values[n]
-            else:
-                del store[n]
-
-    return set_var, cleanup
-
-
-@pytest.fixture(scope="function")
-def env_override():
-    set_var, cleanup = temp_set_var(environ)
-    yield set_var
-    cleanup()
-
-
-@pytest.fixture(scope="function")
-def config_override():
-    set_var, cleanup = temp_set_var(config.overrides)
-    yield set_var
-    cleanup()
 
 
 def test_lookup_default():
